BACK TO CLANG LOL

+++ src/SharedSymbols.cpp
@@ -34,7 +34,7 @@ std::string EncodeBase58(const std::vector<uint8_t>& data) {
         int remainder = 0;
         for (size_t i = 0; i < temp.size(); i++) {
             int num = remainder * 256 + temp[i];
-            temp[i] = num / 58;
+            temp[i] = static_cast<uint8_t>(num / 58);
             remainder = num % 58;
         }
